Add component for Demo login and move code into it
