+++ client/src/pages/Login/Login.tsx
@@ -36,23 +36,6 @@ export default function Login(): JSX.Element {
     });
   };
 
-  const handleDemoLogin = () => {
-    login('jondoe@vmail.com', '12345678').then((data) => {
-      if (data.error) {
-        // setSubmitting(false);
-        updateSnackBarMessage(data.error.message);
-      } else if (data.success) {
-        updateLoginContext(data.success);
-      } else {
-        // should not get here from backend but this catch is for an unknown issue
-        console.error({ data });
-
-        // setSubmitting(false);
-        updateSnackBarMessage('An unexpected error occurred. Please try again');
-      }
-    });
-  };
-
   return (
     <Grid container component="main" className={classes.root}>
       <CssBaseline />
@@ -68,7 +51,7 @@ export default function Login(): JSX.Element {
                   </Typography>
                 </Grid>
               </Grid>
-              <LoginForm handleDemoLogin={handleDemoLogin} handleSubmit={handleSubmit} />
+              <LoginForm handleSubmit={handleSubmit} />
             </Box>
             <Box p={1} alignSelf="center" />
           </Box>
