+++ client/src/pages/Login/LoginForm/LoginForm.tsx
@@ -7,6 +7,7 @@ import Typography from '@material-ui/core/Typography';
 import useStyles from './useStyles';
 import { CircularProgress } from '@material-ui/core';
 import AuthSubText from '../../../components/AuthSubText/AuthSubText';
+import DemoLogin from '../../../components/DemoLogin/DemoLogin';
 
 interface Props {
   handleSubmit: (
@@ -25,11 +26,9 @@ interface Props {
       password: string;
     }>,
   ) => void;
-
-  handleDemoLogin: () => void;
 }
 
-export default function Login({ handleSubmit, handleDemoLogin }: Props): JSX.Element {
+export default function Login({ handleSubmit }: Props): JSX.Element {
   const classes = useStyles();
 
   return (
@@ -94,18 +93,7 @@ export default function Login({ handleSubmit, handleDemoLogin }: Props): JSX.Ele
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
               {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Login'}
             </Button>
-            <Button
-              onClick={() => {
-                handleDemoLogin();
-              }}
-              type="button"
-              size="large"
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-            >
-              {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Demo user'}
-            </Button>
+            <DemoLogin />
             <AuthSubText linkTo="/signup" linkText="Signup" text="Don't have an account?"></AuthSubText>
           </Box>
           <div style={{ height: 95 }} />
