+++ client/src/pages/SignUp/SignUpForm/SignUpForm.tsx
@@ -7,6 +7,7 @@ import Typography from '@material-ui/core/Typography';
 import useStyles from './useStyles';
 import { CircularProgress } from '@material-ui/core';
 import AuthSubText from '../../../components/AuthSubText/AuthSubText';
+import DemoLogin from '../../../components/DemoLogin/DemoLogin';
 
 interface Props {
   handleSubmit: (
@@ -28,11 +29,9 @@ interface Props {
       username: string;
     }>,
   ) => void;
-
-  handleDemoLogin: () => void;
 }
 
-const SignUpForm = ({ handleSubmit, handleDemoLogin }: Props): JSX.Element => {
+const SignUpForm = ({ handleSubmit }: Props): JSX.Element => {
   const classes = useStyles();
 
   return (
@@ -120,18 +119,7 @@ const SignUpForm = ({ handleSubmit, handleDemoLogin }: Props): JSX.Element => {
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
               {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Sign up'}
             </Button>
-            <Button
-              onClick={() => {
-                handleDemoLogin();
-              }}
-              type="button"
-              size="large"
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-            >
-              {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Demo user'}
-            </Button>
+            <DemoLogin />
             <AuthSubText linkTo="/login" linkText="Login" text="Already a member?"></AuthSubText>
           </Box>
         </form>
